% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-data.R
\docType{data}
\name{aquatic_data}
\alias{aquatic_data}
\title{Sample of Aquatic Concentration Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame})
}
\usage{
aquatic_data
}
\description{
Sample data set used for examples.
}
\examples{
head(aquatic_data)
}
\keyword{datasets}
