% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wqbenchdata.R
\name{get_wqbenchdata}
\alias{get_wqbenchdata}
\title{Get wqbench data set}
\usage{
get_wqbenchdata(dataset)
}
\arguments{
\item{dataset}{A data set name}
}
\value{
A data frame
}
\description{
Get wqbench data set
}
\examples{
data <- get_wqbenchdata("aquatic_data")
}
