% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Sample of Raw Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl},
\code{data.frame})
}
\usage{
raw_data
}
\description{
Sample of raw set used for testing.
}
\examples{
head(raw_data)

}
\keyword{datasets}
