% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqbenchdata-package.R
\docType{package}
\name{wqbenchdata-package}
\alias{wqbenchdata}
\alias{wqbenchdata-package}
\title{wqbenchdata: Water Quality Benchmark Data}
\description{
Reference data from the ecotox database used for examples and testing in the wqbench package.
}
\author{
\strong{Maintainer}: Angeline Tillmanns \email{Angeline.Tillmanns@gov.bc.ca}

Authors:
\itemize{
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contributor]
}

}
\keyword{internal}
